Backport 3.2.x] Re-implement thesaurus import/export using rdflib to support a wider range of RDF formats [Fixes #8701] (#8702)
* Use rdflib to load/dump thesauri, instead of bespoke XML parsing

This handles loading RDF in other formats, such as TTL, and other
flavours of RDF+XML.

To handle qualified language tags the length of the DB field
`ThesaurusKeywordLabel.lang` has been increased and a new migration
added.

An additional test has been added for the `dump_thesaurus` command.

Fixes #8701.

* Fix some formatting/pep8 issues

* Adjust length of thesaurus/keyword lang fields to 10 chars

This should allow a few qualifiers in addition to the language code.

* Ensure RDF parser can guess type from file name when uploading thesaurus

When uploaded via the admin an UploadedFile object is provided instead
of a file path; in this case we need to try and guess the RDF format
from the file name ourselves and provide it to the parser.

Also added additional tests for this case.

* Additional pep8 fix

Co-authored-by: mattiagiupponi <51856725+mattiagiupponi@users.noreply.github.com>
Co-authored-by: Alessio Fabiani <alessio.fabiani@geo-solutions.it>